+++ src/main/java/de/samply/fhir/ValidationUtils.java
@@ -8,6 +8,7 @@ import de.samply.template.ConverterTemplate;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.atomic.AtomicReference;
 
 public class ValidationUtils {
@@ -26,7 +27,8 @@ public class ValidationUtils {
     List<String> result = new ArrayList<>();
     validationResult.getMessages().forEach(singleValidationMessage -> {
       if (isMessageForAttributeTemplate(singleValidationMessage, attributeTemplate)) {
-        result.add(singleValidationMessage.getMessage().replace(converterTemplate.getCsvSeparator(), " "));
+        result.add(
+            singleValidationMessage.getMessage().replace(converterTemplate.getCsvSeparator(), " "));
       }
     });
     return result.toArray(new String[0]);
@@ -35,18 +37,58 @@ public class ValidationUtils {
   private static boolean isMessageForAttributeTemplate(
       SingleValidationMessage singleValidationMessage,
       AttributeTemplate attributeTemplate) {
-    String valFhirPath = attributeTemplate.getValFhirPath();
-    String valFhirPath2 = removeSpecialBlocks(valFhirPath);
-    String locationString = singleValidationMessage.getLocationString();
-    String locationString2 = removeSpecialBlocks(locationString);
-    return valFhirPath.contains(locationString)
-        || valFhirPath.contains(locationString2)
-        || valFhirPath2.contains(locationString)
-        || valFhirPath2.contains(locationString2)
-        || locationString.contains(valFhirPath)
-        || locationString.contains(valFhirPath2)
-        || locationString2.contains(valFhirPath)
-        || locationString2.contains(valFhirPath2);
+    boolean result = isMatch(attributeTemplate.getValFhirPath(),
+        singleValidationMessage.getLocationString());
+    if (!result) {
+      return result;
+    }
+    Optional<String> fhirPathInMessage = extractFhirPathInMessage(singleValidationMessage);
+    return fhirPathInMessage.isPresent() ?
+        ValidationUtils.isMatch(attributeTemplate.getValFhirPath(), fhirPathInMessage.get())
+        : true;
+  }
+
+  private static Optional<String> extractFhirPathInMessage(
+      SingleValidationMessage singleValidationMessage) {
+    int index = singleValidationMessage.getLocationString().indexOf(".");
+    String fhirRootElement =
+        (index > 0) ? singleValidationMessage.getLocationString().substring(0, index)
+            : singleValidationMessage.getLocationString();
+    String message = singleValidationMessage.getMessage();
+    for (String part : message.split(" ")) {
+      if (part.contains(fhirRootElement) && part.contains(".")) {
+        int lastIndex = part.lastIndexOf(".");
+        // Remove dot at the end of the string
+        if (lastIndex == part.length() - 1) {
+          part = part.substring(0, lastIndex);
+        }
+        if (part.contains(".")) {
+          part = part.replace(":", "");
+          part = part.replace(",", "");
+          return Optional.of(part);
+        }
+      }
+    }
+    return Optional.empty();
+  }
+
+  private static boolean isMatch(String fhirPath1, String fhirPath2) {
+    String simplifiedFhirPath1 = removeSpecialBlocks(fhirPath1);
+    String simplifiedFhirPath2 = removeSpecialBlocks(fhirPath2);
+    String[] fhirPathGroup1 = {fhirPath1, simplifiedFhirPath1};
+    String[] fhirPathGroup2 = {fhirPath2, simplifiedFhirPath2};
+    return (isMatch(fhirPathGroup1, fhirPathGroup2) || isMatch(fhirPathGroup2, fhirPathGroup1));
+  }
+
+  private static boolean isMatch(String[] fhirPathGroup1, String[] fhirPathGroup2) {
+    for (String fhirPath1 : fhirPathGroup1) {
+      for (String fhirPath2 : fhirPathGroup2) {
+        if (fhirPath1.contains(fhirPath2)) {
+          return true;
+        }
+      }
+    }
+    return false;
   }
 
   private static String removeSpecialBlocks(String input) {